+++ src/main/java/de/mossgrabers/convertwithmoss/format/ni/maschine/maschine2/Sound.java
@@ -1,6 +1,0 @@
-package de.mossgrabers.convertwithmoss.format.ni.maschine.maschine2;
-
-public class Sound
-{
-
-}